etl_app/requirements.txt
Flask
google-cloud-storage
pandas
gunicorn